Move drawItemStack out of hud.h/cpp
