+++ src/gui/CMakeLists.txt
@@ -27,5 +27,6 @@ set(gui_SRCS
 	${CMAKE_CURRENT_SOURCE_DIR}/touchcontrols.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/touchscreenlayout.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/touchscreeneditor.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/drawItemStack.cpp
 	PARENT_SCOPE
 )
