+++ src/gui/guiButtonItemImage.cpp
@@ -5,7 +5,6 @@
 #include "guiButtonItemImage.h"
 
 #include "client/client.h"
-#include "client/hud.h" // drawItemStack
 #include "guiItemImage.h"
 #include "IGUIEnvironment.h"
 #include "itemdef.h"
